test: 更新所有集成测试和前端代码适配CAS架构
全面更新测试套件和前端代码，移除传统格式相关逻辑，适配CAS存储架构。

集成测试更新 (tests/)：

CAS性能测试：
- cas_performance_tests.rs: 更新CAS性能基准测试
- cas_migration_property_tests.rs: 新增CAS迁移属性测试
- cas_test_helpers.rs: 添加CAS测试辅助工具
- cas_helpers_usage_example.rs: CAS使用示例

归档测试：
- archive_manager_integration.rs: 更新归档管理器测试
- archive_processing_integration.rs: 更新归档处理测试
- advanced_features_tests.rs: 新增高级特性测试

错误恢复测试：
- error_recovery_tests.rs: 更新错误恢复测试场景

工作区测试：
- workspace_management_tests.rs: 简化工作区管理测试

搜索测试：
- search_integration_tests.rs: 更新搜索集成测试
- search_property_tests.rs: 更新搜索属性测试

其他测试：
- async_context_property_tests.rs: 更新异步上下文测试
- task_manager_integration_tests.rs: 更新任务管理器测试

核心服务更新 (src-tauri/src/)：

归档模块 (archive/)：
- archive_handler.rs: 更新归档处理器接口
- checkpoint_manager.rs: 更新检查点管理器
- gz_handler.rs: 更新GZ处理器
- parallel_processor.rs: 更新并行处理器
- processor.rs: 更新处理器实现
- public_api.rs: 更新公共API

命令模块 (commands/)：
- config.rs: 新增配置命令
- performance.rs: 更新性能监控命令
- state_sync.rs: 更新状态同步命令
- virtual_tree.rs: 更新虚拟树命令
- legacy.rs: 新增传统格式检测工具
- integration_code.txt: 集成代码参考

模型模块 (models/)：
- config.rs: 更新配置模型
- filters.rs: 更新过滤器模型
- state.rs: 更新状态模型

监控模块 (monitoring/)：
- alerting.rs: 更新告警逻辑
- metrics_collector.rs: 更新指标收集器

搜索引擎 (search_engine/)：
- advanced_features.rs: 更新高级特性
- boolean_query_processor.rs: 更新布尔查询处理器
- concurrent_search.rs: 更新并发搜索
- highlighting_engine.rs: 更新高亮引擎
- index_optimizer.rs: 更新索引优化器
- query_optimizer.rs: 更新查询优化器
- schema.rs: 更新模式定义
- streaming_builder.rs: 更新流式构建器

服务模块 (services/)：
- file_watcher.rs: 更新文件监听器
- index_validator.rs: 更新索引验证器
- metadata_db.rs: 大幅简化元数据数据库，移除传统格式支持
- workspace_metrics.rs: 更新工作区指标

存储模块 (storage/)：
- cas.rs: 更新CAS存储实现
- metadata_store.rs: 更新元数据存储

工具模块 (utils/)：
- cache_manager.rs: 更新缓存管理器
- legacy_detection.rs: 新增传统格式检测工具

其他模块：
- task_manager/mod.rs: 更新任务管理器
- state_sync/mod.rs: 更新状态同步

前端更新 (src/)：

页面组件 (pages/)：
- SettingsPage.tsx: 更新设置页面
- TasksPage.tsx: 更新任务页面
- WorkspacesPage.tsx: 简化工作区页面，移除迁移UI

组件 (components/)：
- ErrorFallback.tsx: 更新错误边界

Hooks：
- index.ts: 移除useMigration导出

状态管理 (stores/)：
- workspaceStore.ts: 更新工作区状态管理

配置：
- setupTests.ts: 更新测试配置

其他：
- Cargo.lock: 更新依赖锁定
- tauri.conf.json: 更新Tauri配置
- migrations/README.md: 更新迁移文档说明

统计：
- 66个文件修改
- 新增4457行代码
- 删除1807行代码
- 净增2650行代码

影响：
- 所有测试用例适配CAS架构
- 前端UI移除迁移相关功能
- 核心服务大幅简化
- 测试覆盖率保持

+++ log-analyzer/src-tauri/src/search_engine/boolean_query_processor.rs
@@ -1,4 +1,5 @@
 //! Boolean Query Processor
+#![allow(dead_code)]
 //!
 //! Optimized multi-keyword intersection algorithms using Tantivy's boolean query capabilities.
 //! Provides:
@@ -164,7 +165,7 @@ impl BooleanQueryProcessor {
 
         let count_collector = Count;
         let doc_count = searcher.search(&term_query, &count_collector)?;
-        let total_docs = searcher.num_docs() as u64;
+        let total_docs = searcher.num_docs();
 
         let selectivity = if total_docs > 0 {
             doc_count as f64 / total_docs as f64
